Add files via upload
there are two main python codes which are essential. first is captureimage.py which is used to capture image using webcam(run the program, press space to click image and esc to quit the program)
second is facedetect.py in which you set the path(name of image file you want to detect face on)  then run it simply. it will give you an image with rectangle in faces detected.